Fix for Wndows compilation
