+++ lib/Support/Time.cpp
@@ -13,8 +13,6 @@
 #include <sstream>
 #include <tuple>
 
-#define RUSAGE_SELF 0
-
 #ifdef _WIN32
 #include <psapi.h>
 #include <windows.h>
@@ -22,6 +20,45 @@
 #include <sys/resource.h>
 #include <sys/time.h>
 #endif
+
+#ifdef _WIN32
+// Windows-specific implementation of getrusage
+int getrusage(int who, struct rusage *usage) {
+
+  FILETIME creation_time, exit_time, kernel_time, user_time;
+  PROCESS_MEMORY_COUNTERS pmc;
+
+  if (who != RUSAGE_SELF) {
+    return -1; // Only RUSAGE_SELF is supported on Windows
+  }
+
+  if (!GetProcessTimes(GetCurrentProcess(), &creation_time, &exit_time,
+                       &kernel_time, &user_time)) {
+    return -1;
+  }
+
+  if (!GetProcessMemoryInfo(GetCurrentProcess(), &pmc, sizeof(pmc))) {
+    return -1;
+  }
+
+  // Convert FILETIME to timeval
+  ULARGE_INTEGER kernel, user;
+  kernel.LowPart = kernel_time.dwLowDateTime;
+  kernel.HighPart = kernel_time.dwHighDateTime;
+  user.LowPart = user_time.dwLowDateTime;
+  user.HighPart = user_time.dwHighDateTime;
+
+  usage->ru_utime.tv_sec = (long)(user.QuadPart / 10000000);
+  usage->ru_utime.tv_usec = (long)((user.QuadPart % 10000000) / 10);
+  usage->ru_stime.tv_sec = (long)(kernel.QuadPart / 10000000);
+  usage->ru_stime.tv_usec = (long)((kernel.QuadPart % 10000000) / 10);
+
+  usage->ru_maxrss =
+      (long)(pmc.PeakWorkingSetSize / 1024); // Convert to kilobytes
+
+  return 0;
+}
+#endif 
 
 using namespace klee;
